requirements-optimized.txt
# Optimized dependencies for fast, accurate Bhagavad Gita search
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# PDF processing
pymupdf>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0

# Vector Database (optimized)
faiss-cpu>=1.8.0

# NLP and text processing
numpy>=1.24.0
language-tool-python>=1.0.0
nltk>=3.8.1
psutil>=5.9.0

# Machine Learning and Embeddings (optimized versions)
torch>=2.0.0
sentence-transformers>=3.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# Fuzzy matching and string similarity
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0  # C acceleration for fuzzy matching

# Text processing and utilities
huggingface-hub>=0.16.0
requests>=2.28.0
tqdm>=4.64.0

# LangChain for text splitting
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# LLM Integration
openai>=1.0.0

# Google Cloud for embeddings
google-generativeai>=0.3.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0

# Firestore and Firebase for chat history
google-cloud-firestore>=2.14.0
firebase-admin>=6.4.0

# Authentication and security
pyjwt>=2.8.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1

# Additional utilities
python-dateutil>=2.8.2
flake8>=6.1.0
mypy>=1.5.1
